fix(v4): la familia ESP32 llevaba sin compilar — y el S3 mentía sobre su ISA
Al ir a construir el P4 para #311 saltaron tres cosas que estaban rotas desde
antes, ninguna detectada porque no se había reconstruido la familia:

1. `events.c` (H5.c) nunca entró en los CMakeLists del S3 ni del P4 → el enlace
   moría con cuatro símbolos sin resolver (queue_init, drain_one, mark_roots,
   enqueue). Los eventos existían en el core pero ninguna placa ESP32 los tenía.

2. `events.c` imprimía `tc->id` (int32_t) con `%d`. En RISC-V int32_t es `long`,
   así que con -Werror no pasa. Casteado como los otros dos fprintf del fichero.

3. `repl_esp32.c` publica `arch` en el INFO (H11: la placa dice a qué ISA
   compilar el nativo) pero incluía `mdn_loader.h` DENTRO del `#if __riscv`, que
   es sólo para la carga de .mdn. El S3 se quedaba sin la declaración.

Y al enlazar `mdn_loader.c` en el S3 se destapó lo de verdad importante, que
cazó Eduardo: **en Xtensa el mapa de arquitecturas no tenía rama**, así que caía
en el `else` = MDN_ARCH_NONE. Eso NO significa "Xtensa", significa "host, SIN
gate", con daño doble: el IDE recibía arch=0, no sabía a qué compilar y tiraba
del ajuste del proyecto (que puede decir "arm"), y además el gate de
arquitectura del loader se DESARMA (`if (MDN_HOST_ARCH != MDN_ARCH_NONE)`) → un
.mdn de otra ISA se aceptaría. Hoy inalcanzable porque el S3 no carga .mdn, pero
enlazar el fichero lo ponía a tiro: misma familia que el ABI rancio de 76222a9.

Se le enseña Xtensa (EM_XTENSA=94) en vez de ampliar el "sin gate": el S3 dice
la verdad y AotBuild:126 lo rechaza con un aviso claro ("target no soportado")
en lugar de colarle código ARM a una placa Xtensa. Fallar limpio, no en silencio.

Verificado: compilan P4, S3 y host.

+++ bpgenvm-c/src/events.c
@@ -114,7 +114,8 @@ static int inject(bpvm_t* vm, bpvm_thread_t* tc, const bpvm_event_t* e) {
     uint32_t need = 4u + BPVM_REF_SIZE + 12u;
     for (int i = 0; i < e->nargs; i++) need += (e->masks & (1u << (8 + i))) ? 8u : 4u;
     if (tc->sp + need > tc->stack_top) {
-        fprintf(stderr, "[bpvm] evento: sin pila en tid=%d — descartado\n", tc->id);
+        /* cast: int32_t no es `int` en todos los ports (en RISC-V es `long`) */
+        fprintf(stderr, "[bpvm] evento: sin pila en tid=%d — descartado\n", (int) tc->id);
         return 0;
     }
